Add Merit Removals to the State.
Since reversions requires loading old Blocks, this also adds Element parsing to Blocks.

Also fixes a bug in the Blockchain's checking that if a MeritRemoval is archived, it's the only Element archived.

Also corrects handling of the contents Merkle.

+++ src/Database/Filesystem/DB/objects/DBObj.nim
@@ -22,7 +22,7 @@ type
 
     MeritDB* = ref object
         cache*: Table[string, string]
-        removals*: Table[int, int]
+        removals*: Table[uint16, int]
         holders*: string
 
     DB* = ref object
@@ -47,7 +47,7 @@ proc newConsensusDB(): ConsensusDB {.inline, forceCheck: [].} =
 proc newMeritDB(): MeritDB {.inline, forceCheck: [].} =
     MeritDB(
         cache: initTable[string, string](),
-        removals: initTable[int, int](),
+        removals: initTable[uint16, int](),
         holders: ""
     )
